Various OG image improvements
Some better naming and switching from controller to plug, that just feels a lot better.
