+++ lib/jola_dev/og_image.ex
@@ -1,21 +1,22 @@
 defmodule JolaDev.OGImage do
   @moduledoc """
   Per-page Open Graph image lookup. All images are rendered at compile time
-  via `JolaDev.OGImage.Renderer` and baked into the `@baked` module
-  attribute, then served at runtime from `JolaDevWeb.OGImageController`.
+  via `JolaDev.OGImage.Renderer` and baked into the `@images` module
+  attribute, then served at runtime from `JolaDevWeb.Plugs.OGImage`.
 
   See `JolaDev.OGImage.Renderer` for the rendering primitives.
   """
 
+  alias JolaDev.OGImage.Catalog
   alias JolaDev.OGImage.Renderer
 
   @dev_mode Application.compile_env!(:jola_dev, :og_image_dev_mode)
-  @baked if not @dev_mode,
-           do:
-             Map.new(Renderer.all_slugs(), fn slug ->
-               {title, description} = Renderer.content_for(slug)
-               {slug, Renderer.generate_bytes(title, description)}
-             end)
+  @images if not @dev_mode,
+            do:
+              Map.new(Catalog.all_slugs(), fn slug ->
+                {title, description} = Catalog.content_for(slug)
+                {slug, Renderer.generate_bytes(title, description)}
+              end)
 
   @doc """
   Returns the public asset path for a slug's OG image. Used by controllers,
@@ -26,16 +27,12 @@ defmodule JolaDev.OGImage do
   @doc """
   Returns `{:ok, png_bytes}` for a known slug, `:error` otherwise.
   """
-  def bytes_for(slug) do
-    if @baked do
-      Map.fetch(@baked, slug)
+  def image_for(slug) do
+    if @images do
+      Map.fetch(@images, slug)
     else
-      case Renderer.content_for(slug) do
-        {title, description} ->
-          {:ok, Renderer.generate_bytes(title, description)}
-
-        :error ->
-          :error
+      with {title, description} <- Catalog.content_for(slug) do
+        {:ok, Renderer.generate_bytes(title, description)}
       end
     end
   end
